Put together box types in single source file
