--- elst_test.go
@@ -1,100 +1,0 @@
-package mp4
-
-import (
-	"bytes"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-)
-
-func TestElstMarshal(t *testing.T) {
-	testCases := []struct {
-		name string
-		elst Elst
-		want []byte
-	}{
-		{
-			name: "version 0",
-			elst: Elst{
-				FullBox: FullBox{
-					Version: 0,
-					Flags:   [3]byte{0x00, 0x00, 0x00},
-				},
-				EntryCount: 2,
-				Entries: []ElstEntry{
-					{
-						SegmentDurationV0: 0x0100000a,
-						MediaTimeV0:       0x0100000b,
-						MediaRateInteger:  0x010c,
-						MediaRateFraction: 0x010d,
-					}, {
-						SegmentDurationV0: 0x0200000a,
-						MediaTimeV0:       0x0200000b,
-						MediaRateInteger:  0x020c,
-						MediaRateFraction: 0x020d,
-					},
-				},
-			},
-			want: []byte{
-				0,                // version
-				0x00, 0x00, 0x00, // flags
-				0x00, 0x00, 0x00, 0x02, // entry count
-				0x01, 0x00, 0x00, 0x0a, // segment duration v0
-				0x01, 0x00, 0x00, 0x0b, // media time v0
-				0x01, 0x0c, // media rate integer
-				0x01, 0x0d, // media rate fraction
-				0x02, 0x00, 0x00, 0x0a, // segment duration v0
-				0x02, 0x00, 0x00, 0x0b, // media time v0
-				0x02, 0x0c, // media rate integer
-				0x02, 0x0d, // media rate fraction
-			},
-		},
-		{
-			name: "version 1",
-			elst: Elst{
-				FullBox: FullBox{
-					Version: 1,
-					Flags:   [3]byte{0x00, 0x00, 0x00},
-				},
-				EntryCount: 2,
-				Entries: []ElstEntry{
-					{
-						SegmentDurationV1: 0x010000000000000a,
-						MediaTimeV1:       0x010000000000000b,
-						MediaRateInteger:  0x010c,
-						MediaRateFraction: 0x010d,
-					}, {
-						SegmentDurationV1: 0x020000000000000a,
-						MediaTimeV1:       0x020000000000000b,
-						MediaRateInteger:  0x020c,
-						MediaRateFraction: 0x020d,
-					},
-				},
-			},
-			want: []byte{
-				1,                // version
-				0x00, 0x00, 0x00, // flags
-				0x00, 0x00, 0x00, 0x02, // entry count
-				0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0a, // segment duration v1
-				0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0b, // media time v1
-				0x01, 0x0c, // media rate integer
-				0x01, 0x0d, // media rate fraction
-				0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0a, // segment duration v1
-				0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0b, // media time v1
-				0x02, 0x0c, // media rate integer
-				0x02, 0x0d, // media rate fraction
-			},
-		},
-	}
-
-	for _, c := range testCases {
-		t.Run(c.name, func(t *testing.T) {
-			buf := bytes.NewBuffer(nil)
-			n, err := Marshal(buf, &c.elst)
-			require.NoError(t, err)
-			assert.Equal(t, uint64(len(c.want)), n)
-			assert.Equal(t, c.want, buf.Bytes())
-		})
-	}
-}
